refactor(utils): add isMatrix and extend
