+++ src/utils/index.ts
@@ -1,1 +1,3 @@
 export { compose } from './compose';
+export { isMatrix } from './isMatrix';
+export { extend } from './extend';
