+++ src/utils/isMatrix.ts
@@ -1,0 +1,4 @@
+import { Matrix3 } from '../type';
+export function isMatrix(transformer: any): transformer is Matrix3 {
+  return transformer instanceof Float32Array || transformer instanceof Array;
+}
